【GEC-SF】CAN: 1、解决接收数据不完整造成错误的问题 2、发送数据时判断是否有空邮箱发送

+++ HARDWARE/CAN/can.h
@@ -18,14 +18,14 @@
 typedef struct  {
   u8 sending;
   u8* p_CanBuff;
-  u8 mlen;
+  s16 mlen;
   u8 tx_buff[200];
 } CAN_TX_DATA_PROCESS_TypeDef;
 
 typedef struct  {
   u8 recving;
-  u8 rxcnt;
-  u8 mlen;
+  s16 rxcnt;
+  s16 mlen;
   u8 data_packet;
   u8 recv_cnt;
   u8 recv_len;
